Feat: adds submitted state representation on survey
