+++ package.json
@@ -52,7 +52,7 @@
     "react-dom": "^18",
     "react-hook-form": "^7.56.3",
     "sass": "^1.77.8",
-    "styled-components": "^6.1.8",
+    "styled-components": "^6.1.19",
     "swiper": "^11.1.14",
     "tailwind-merge": "^3.2.0",
     "ua-parser-js": "^2.0.4",
